pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["flows", "compute"]

[project]
name = "globus-workflow"
version = "0.1.0"
description = "Globus workflow components and scripts for SYNAPS-I"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "BSD-3-Clause"}

dependencies = [
    "globus-sdk>=3.0,<4",
    "globus-compute-sdk>=2.0",
    "pyyaml>=6.0",
    "watchdog>=3.0",
    "wandb>=0.15",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]
ignore = [
    "E501", # line-too-long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["utils"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
